+++ web_google_maps/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add toggle button (show/hide) google_map sidebar
 - Add new button (custom control) in the map, a button to center map
 ### Changed
+- Updated map gestureHandling value, added all supported values `auto`, `greedy`, `cooperative`, and `none` (previous only two: `greedy` and `cooperative`). For more detail check https://developers.google.com/maps/documentation/javascript/interaction#controlling_gesture_handling
 ### Fixed
 
 ## [14.0.2.0.5] - 2021-08-29